WIP: buildFilesToProcess in ValidationRunner
